requirements.txt
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
djangorestframework==3.11.0
fpdf==1.7.2
gunicorn==20.0.0
idna==2.9
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==4.1.4
